Move the most function declarations from fping.c to fping.h
